refactored all UX modules
