+++ liquidhaskell.cabal
@@ -137,7 +137,7 @@ Library
    Exposed-Modules: Language.Haskell.Liquid.Prelude,
                     Language.Haskell.Liquid.Foreign,
                     Language.Haskell.Liquid.List,
-                    Language.Haskell.Liquid.PrettyPrint,
+                    Language.Haskell.Liquid.UX.PrettyPrint,
                     Language.Haskell.Liquid.Bare,
                     Language.Haskell.Liquid.Constraint.Constraint,
                     Language.Haskell.Liquid.Constraint.Types,
@@ -150,16 +150,16 @@ Library
                     Language.Haskell.Liquid.GhcInterface,
                     Language.Haskell.Liquid.Types.RefType,
                     Language.Haskell.Liquid.RefSplit,
-                    Language.Haskell.Liquid.Errors,
+                    Language.Haskell.Liquid.UX.Errors,
                     Language.Haskell.Liquid.Types.PredType,
-                    Language.Haskell.Liquid.ACSS,
-                    Language.Haskell.Liquid.DiffCheck,
+                    Language.Haskell.Liquid.UX.ACSS,
+                    Language.Haskell.Liquid.UX.DiffCheck,
                     Language.Haskell.Liquid.ANFTransform,
-                    Language.Haskell.Liquid.Annotate,
-                    Language.Haskell.Liquid.CTags,
-                    Language.Haskell.Liquid.Config,
-                    Language.Haskell.Liquid.CmdLine,
-                    Language.Haskell.Liquid.Cabal,
+                    Language.Haskell.Liquid.UX.Annotate,
+                    Language.Haskell.Liquid.UX.CTags,
+                    Language.Haskell.Liquid.UX.Config,
+                    Language.Haskell.Liquid.UX.CmdLine,
+                    Language.Haskell.Liquid.UX.Cabal,
                     Language.Haskell.Liquid.GhcMisc,
                     Language.Haskell.Liquid.GhcPlay,
                     Language.Haskell.Liquid.Misc,
@@ -169,7 +169,7 @@ Library
                     Language.Haskell.Liquid.Types.Dictionaries,
                     Language.Haskell.Liquid.Constraint.Qualifier,
                     Language.Haskell.Liquid.TransformRec,
-                    Language.Haskell.Liquid.Tidy,
+                    Language.Haskell.Liquid.UX.Tidy,
                     Language.Haskell.Liquid.Types,
                     Language.Haskell.Liquid.Simplify,
                     Language.Haskell.Liquid.Types.Literals,
