+++ src/Language/Haskell/Liquid/UX/Annotate.hs
@@ -12,7 +12,7 @@
 -- 3. JSON files for the web-demo etc.
 ---------------------------------------------------------------------------
 
-module Language.Haskell.Liquid.Annotate (mkOutput, annotate) where
+module Language.Haskell.Liquid.UX.Annotate (mkOutput, annotate) where
 
 import           GHC                      ( SrcSpan (..)
                                           , srcSpanStartCol
@@ -41,17 +41,17 @@ import qualified Data.Vector            as V
 import qualified Data.ByteString.Lazy   as B
 import qualified Data.Text              as T
 import qualified Data.HashMap.Strict    as M
-import qualified Language.Haskell.Liquid.ACSS as ACSS
+import qualified Language.Haskell.Liquid.UX.ACSS as ACSS
 import           Language.Haskell.HsColour.Classify
 import           Language.Fixpoint.Utils.Files
 import           Language.Fixpoint.Misc
 import           Language.Haskell.Liquid.GhcMisc
 import           Language.Fixpoint.Types hiding (Error, Loc, Def (..), Constant (..), Located (..))
 import           Language.Haskell.Liquid.Misc
-import           Language.Haskell.Liquid.PrettyPrint
+import           Language.Haskell.Liquid.UX.PrettyPrint
 import           Language.Haskell.Liquid.Types.RefType
-import           Language.Haskell.Liquid.Errors ()
-import           Language.Haskell.Liquid.Tidy
+import           Language.Haskell.Liquid.UX.Errors ()
+import           Language.Haskell.Liquid.UX.Tidy
 import           Language.Haskell.Liquid.Types hiding (Located(..), Def(..))
 
 -- | @output@ creates the pretty printed output
