+++ src/Language/Haskell/Liquid/UX/Cabal.hs
@@ -11,7 +11,7 @@
 {-# LANGUAGE OverloadedStrings    #-}
 {-# LANGUAGE FlexibleInstances    #-}
 
-module Language.Haskell.Liquid.Cabal (cabalInfo, Info(..)) where
+module Language.Haskell.Liquid.UX.Cabal (cabalInfo, Info(..)) where
 
 import Control.Applicative ((<$>))
 import Data.Bits                              ( shiftL, shiftR, xor )
@@ -37,7 +37,7 @@ import System.Exit
 import System.FilePath
 import System.Directory
 import System.Info
-import Language.Haskell.Liquid.Errors
+import Language.Haskell.Liquid.UX.Errors
 
 -- To use in ghci:
 --   exitWithPanic = undefined
