+++ src/Language/Haskell/Liquid/UX/DiffCheck.hs
@@ -7,7 +7,7 @@
 {-# LANGUAGE FlexibleContexts          #-}
 {-# LANGUAGE FlexibleInstances         #-}
 
-module Language.Haskell.Liquid.DiffCheck (
+module Language.Haskell.Liquid.UX.DiffCheck (
 
    -- * Changed binders + Unchanged Errors
      DiffCheck (..)
@@ -47,7 +47,7 @@ import            Language.Haskell.Liquid.Types   (SpecType, GhcSpec (..), AnnIn
 import            Language.Haskell.Liquid.Misc    (mkGraph)
 import            Language.Haskell.Liquid.GhcMisc
 import            Language.Haskell.Liquid.Types.Visitors
-import            Language.Haskell.Liquid.Errors   ()
+import            Language.Haskell.Liquid.UX.Errors   ()
 import            Text.Parsec.Pos                  (sourceName, sourceLine, sourceColumn, SourcePos, newPos)
 import            Text.PrettyPrint.HughesPJ        (text, render, Doc)
